Detect unexpected log rotation and release renamed files
In the previous vesions, there was no way to detect log rotation
event when log file was rotated by external tool (logrotate)

It causes that rotated log file was grabbed by DaemonLogger.
This was unexpected behavior.

This commit fixes by reopening log file correctly.

NOTE: This PR depends on inotify so only linux was supported.
Usually Fluentd built-in rotation works so this feature is
optional.

Signed-off-by: Kentaro Hayashi <hayashi@clear-code.com>