Reopen log file when rotation is detected
In the previous vesions, there was no way to detect log rotation event when
log file was rotated by external tool (logrotate)

It causes that DaemonLogger continues to write logs to rotated file even even
when it's already renamed or removed, it's not expected behavior.

This commit fixes by reopening log file when inode is changed.

Signed-off-by: Takuro Ashie <ashie@clear-code.com>