SocketManager: Enable open(nil) selecting path automatically
This `open(nil)` selects `path` automatically.

Especially on Windows, it is hard to search a available port for `path`,
so we should use this feature.
(The current `generate_path` logic sometimes chooses unavailable port.
For example, it does not consider the excluded port range.)

Note: This fix keeps backward compatibility.

Signed-off-by: Daijiro Fukuda <fukuda@clear-code.com>